requirements.txt
python-telegram-bot==3.4
requests==2.9.1